Ruleset: don't fail if Landlock is not supported
The library is designed to do Landlock on a best-effort basis, so
failing if no landlock support is present at all is a bad choice.
Instead, this commit now fails silently if no Landlock support is
available and provides a function to manually check whether support is
available, in case a library consumer needs to know.