Ruleset: add method to get the effective ABI version
We can make use of the effective ABI version when testing for the
correct funciton of the library, as we need to make sure we test
enforcement of rules only when it is actually possible to enforce them.
As this requires both the probed ABI version and the build-time ABI
version to support the rules under test, the minimum of both is a
suitable criterion for whether a rule is enforcable or not.